Add gradle and transfer task methods to TaskList
