+++ src/main/java/bytetalker/parser/Parser.java
@@ -1,5 +1,7 @@
 package bytetalker.parser;
 
+import bytetalker.ui.Ui;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
